added simple transmitter and simple receiver code.
These were used to test the send/receive capabilities of a pair of
transmitter/receivers, which are available from sparkfun.  Now that they
are working, they will serve as a base example to test connections if need
be, and as a solid code example for building wireless signalling between
units.